\name{CEGbase-package}
\alias{CEGbase-package}
\docType{package}
\title{
A collection of utility functions for CEG
}
\description{
This package should contain basic functions necessary for almost all
projects you develop in R.
}
\details{
\tabular{ll}{
Package: \tab CEGbase\cr
Type: \tab Package\cr
Version: \tab 0.0.x\cr
Date: \tab 2009-01-22\cr
License: \tab CEG\cr
}

}
\author{
John Scillieri, Adrian Dragulescu.
}
\references{
}
\keyword{ package }
\examples{
}
